doc: rename repository and program name
