+++ README.md
@@ -1,4 +1,4 @@
-# Calculator
+# Calc
 A calculator console application that allows doing four basic mathematical operations with real numbers
 
 Each calculation involves two numbers only, but after the calculation, the user can choose to save the result and use it for further calculations.
